+++ libraries/ADG2128_RT/ADG2128.h
@@ -3,7 +3,7 @@
 //    FILE: ADG2128.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2025-02-28
-// VERSION: 0.1.0
+// VERSION: 0.2.0
 // PURPOSE: Arduino library for ADG2128 8x12 (cross-point) matrix switch with I2C.
 //     URL: https://github.com/RobTillaart/ADG2128
 //
@@ -13,16 +13,17 @@
 #include "Wire.h"
 
 
-#define ADG2128_LIB_VERSION             (F("0.1.0"))
+#define ADG2128_LIB_VERSION             (F("0.2.0"))
 
 #define ADG2128_DEFAULT_ADDRESS         0x70
 
 //  ERROR CODES
 //  values <> 0 are errors.
 #define ADG2128_OK                      0x00
-#define ADG2128_CRC_ERROR               0x01
-#define ADG2128_NOT_READY               0x10
-#define ADG2128_REQUEST_ERROR           0x11
+#define ADG2128_CRC_ERROR               0x80
+#define ADG2128_NOT_READY               0x81
+#define ADG2128_REQUEST_ERROR           0x82
+#define ADG2128_ADDRESS_ERROR           0x83
 
 
 class ADG2128
@@ -36,20 +37,25 @@ public:
 
 
   //  SWITCH
-  void     on(uint8_t row, uint8_t col);
-  void     off(uint8_t row, uint8_t col);
-  bool     isOn(uint8_t row, uint8_t col);
-  uint8_t  isOnMask(uint8_t col);  //  get a whole column at once as bit mask.
+  //  row == X = 0..11, column == Y = 0..7
+  bool     on(uint8_t row, uint8_t column);
+  bool     off(uint8_t row, uint8_t column);
+  bool     isOn(uint8_t row, uint8_t column);
+  //  get multi status in a bit mask
+  uint8_t  isOnRow(uint8_t row);        //  get a whole row at once as bit mask.
+  //  uint16_t isOnColumn(uint8_t column);  //  get a whole column at once as bit mask.
 
-  //  WRAPPERS.
-  void     on(uint8_t sw)   { on(sw / 12, sw % 12); };
-  void     off(uint8_t sw)  { off(sw / 12, sw % 12); };
-  uint8_t  isOn(uint8_t sw) { return isOn(sw / 12, sw % 12); };
+  //  WRAPPERS
+  //  e.g  switch 29 ==> row (X) = 3  column (Y) = 5
+  bool     on(uint8_t sw)   { return on(sw / 8, sw % 8); };
+  bool     off(uint8_t sw)  { return off(sw / 8, sw % 8); };
+  uint8_t  isOn(uint8_t sw) { return isOn(sw / 8, sw % 8); };
 
 
   //  MODE
   //  default direct (transparent) mode
-  void     setMode(bool latched = false);
+  void     setDirectMode();
+  void     setLatchMode();
   bool     isLatchedMode();
   bool     isDirectMode();
 
@@ -65,11 +71,11 @@ public:
 
 
 private:
-  uint8_t  _address = 0x2A;
+  uint8_t  _address;
   TwoWire* _wire;
 
   int      _send(uint8_t pins, uint8_t value);
-  int      _readback(uint8_t value);
+  int      _readback(uint8_t row);
 
   int      _error;
   uint8_t  _mode;