+++ libraries/ADG2128_RT/CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2025-04-03
+- fix #2, swap rows and columns to match figure 1 datasheet.
+- add **isOnRow(row)** to replace isOnMask()
+- removed **isOnMask()**
+- add bool return value for **bool on()** and **bool off()**
+- replaced setMode() to **setLatchMode()** and **setDirectMode()**
+- renumber error constants
+- update unit test
+- update readme.md
+- update keywords.txt
+- minor edits
+
 ## [0.1.0] - 2025-02-28
 - initial version